% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrich_heatmap.R
\name{gene_enrichment}
\alias{gene_enrichment}
\title{Perform Gene Ontology Enrichments for Multiple Tissue/Cell Types}
\usage{
gene_enrichment(
  gene_set,
  organism = "org.Hs.eg.db",
  tissue_types = names(gene_set),
  background = NULL,
  p_adj_method = "fdr",
  min_gene_set_size = 3,
  max_gene_set_size = 800,
  p_cutoff = 0.05,
  ont_type = "BP",
  gene_nom = "ENSEMBL"
)
}
\arguments{
\item{gene_set}{List where each entry is a set of significant genes for
a tissue.}

\item{organism}{Genome wide annotation for your organsim of choice.
You must have your chosen annotation package installed.}

\item{tissue_types}{Vector of names that correspond to entries of gene_set.}

\item{background}{List where each entry is a vector of gene names that
describe the transcriptomic universe of the corresponding
tissue type in gene_set.}

\item{p_adj_method}{p-value adjustment method. One of "holm", "hochberg",
"hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{min_gene_set_size}{Integer describing the smallest GO term size to be
considered by enrichGO.}

\item{max_gene_set_size}{Integer describing the largest GO term size to be
considered by enrichGO.}

\item{p_cutoff}{Max p-value that will be considered as significant.}

\item{ont_type}{Which GO subontology to query.
One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{gene_nom}{Nomenclature/keytype of genes in gene_set inputs.}
}
\value{
Returns a list of enrichGO() results.
}
\description{
Uses clusterProfiler function enrichGO() to find Gene Ontology (GO)
enrichments for multiple tissue types.
}
\examples{
gene_enrichment(gene_set = list(a=c("CHST2", "B3GNT1", "B3GNT2", "CHST1",
"B4GALT1","B3GNT7", "ST3GAL2", "B4GALT3", "ST3GAL1", "B4GALT2", "B4GALT4",
"CHST4", "ST3GAL3", "FUT8", "CHST6")),
organism = "org.Hs.eg.db", gene_nom = "SYMBOL")

}
\references{
G Yu. Thirteen years of clusterProfiler. The Innovation. 2024, 5(6):100722
}
