% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expr_heatmap_plot.R
\name{reorder_corr_hclust}
\alias{reorder_corr_hclust}
\title{Reorder Correlation Matrix using Complete Hierarchical Clustering}
\usage{
reorder_corr_hclust(corr_mat)
}
\arguments{
\item{corr_mat}{Correlation matrix.}
}
\value{
Returns a reordered correlation matrix.
}
\description{
Produces an a matrix reordered based on complete hierarchical clustering.
}
