% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TWAS_analysis.R
\name{run_TWAS_analysis}
\alias{run_TWAS_analysis}
\title{Launch Shiny App for TWASviz}
\usage{
run_TWAS_analysis()
}
\value{
No return value.
}
\description{
Launches the Shiny app for TWASviz, providing
a user interface to interactively produce volcano plots, overlap heatmaps,
and Gene Ontology term enrichment heatmaps this package.
The app's code is located in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
  run_TWAS_analysis()
}

}
\references{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Xie Y, Allen J, McPherson J, Dipert A, Borges B (2025). \emph{shiny: Web Application Framework for R}. R package version 1.11.1, \url{https://CRAN.R-project.org/package=shiny}.
}
