% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expr_heatmap_plot.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{Heatmap of all Pairwise Correlations of Expression of Many Genes}
\usage{
corr_heatmap(
  gene_expr,
  target_height = 300,
  target_width = 300,
  summary_func = function(x) x[which.max(abs(x))],
  n_core = parallel::detectCores() - 1,
  col_pal = (grDevices::colorRampPalette(c("red", "white", "blue")))(50),
  title = "Gene expression correlation matrix"
)
}
\arguments{
\item{gene_expr}{Gene expression matrix. Rows are individual samples.
Columns are genes.}

\item{target_height}{Max number of rows that the heatmap should have.}

\item{target_width}{Max number of columns that the heatmap should have.}

\item{summary_func}{Function with which to down sample a large
correlation matrix.}

\item{n_core}{Number of cores for parallel processing}

\item{col_pal}{Vector of hex colors, in order of corresponding colours
from -1 to 1.}

\item{title}{Title for the heatmap.}
}
\value{
Returns a heatmap image of gene correlations.
}
\description{
Produces an image of a heatmap for all of the pairwise correlations of
many gene expressions. Intended for use for expression of > 500 genes.
}
\examples{
\dontrun{
  corr_heatmap(gene_expr = matrix(rnorm(500^2), nrow = 500))
}

}
\references{
R Core Team (2024). R: A Language and Environment for Statistical
Computing. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.
}
