% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_overlap_heatmap.R
\name{correlation_overlap_heatmap}
\alias{correlation_overlap_heatmap}
\title{Plot Pairwise Correlations and Overlaps of Multiple Gene Sets}
\usage{
correlation_overlap_heatmap(
  betas,
  cor_method = "pearson",
  tissue_names = 1:ncol(betas)
)
}
\arguments{
\item{betas}{A dataframe with rows of genes and columns of tissue/cell types.
Entry (i, j) is the TWAS effect size of gene i in tissue j, or NULL if gene i
is not expressed/significant in tissue j.}

\item{cor_method}{A string specifying the correlation method
(pearson, kendall, spearman).}

\item{tissue_names}{A vector of strings with length of the number of columns
in betas. These are the tissue names in the order that they appear in betas.}
}
\value{
Returns a heatmap plot of correlations and overlaps
}
\description{
Plots heatmap for pairwise complete correlations for effect sizes of
multiple gene sets in lower triangle, number of pairwise complete effect
sizes for multiple gene sets in upper triangle. Diagonal displays total
number of effect sizes for a single gene set.
}
\examples{
correlation_overlap_heatmap(betas = data.frame(x = c(1, 7, 3), y = c(1, -9, 3),
                            z = c(-1, 0, NA)), tissue_names = c("A", "B", "C"))

}
