% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrich_heatmap.R
\name{goenrich_heatmap}
\alias{goenrich_heatmap}
\title{Plot Gene Ontology Enrichments for Multiple Tissue/Cell Types}
\usage{
goenrich_heatmap(enrich_res, top_n = 5, x_label = "Cell Type")
}
\arguments{
\item{enrich_res}{List where each entry is enrichGO result for a tissue.}

\item{top_n}{Number of GO terms to use per tissue type.}
}
\value{
Returns a heatmap plot of p-value strength of each tissue's
enrichment for top GO terms.
}
\description{
Plots the -log10(adjusted p-value) of the Gene Ontology (GO) enrichments
for multiple tissue types.
}
\examples{
goenrich_heatmap(enrich_res = gene_enrichment(gene_set = list(a=c("CHST2",
"B3GNT1", "B3GNT2", "CHST1",
"B4GALT1","B3GNT7", "ST3GAL2", "B4GALT3", "ST3GAL1", "B4GALT2", "B4GALT4",
"CHST4", "ST3GAL3", "FUT8", "CHST6")),
organism = "org.Hs.eg.db", gene_nom = "SYMBOL"))

}
