% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAS_to_adj_dataframe.R
\name{predixcan2adj_df}
\alias{predixcan2adj_df}
\title{Format Coefficient Output from PrediXcan Association Output.}
\usage{
predixcan2adj_df(
  predixcan_assoc_filenames,
  tissue_names = seq_along(predixcan_assoc_filenames),
  use_fdr = FALSE,
  pvalue_thresh = 0.05
)
}
\arguments{
\item{predixcan_assoc_filenames}{List of paths to file output from
PrediXcan association software.}

\item{tissue_names}{Names for the tissues/cell/condition types for the
imputed gene expression data used to produce each of the above PrediXcan
associations.}
}
\value{
Returns an adjacency-matrix-like dataframe of genes and their
effect sizes for multiple tissues/cells/conditions.
}
\description{
Produces an adjacency-matrix-like dataframe where each row is a gene and
each column is the estimated effect size of the gene from the PrediXcan
association software.
}
