% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_preprocessing.R
\name{preprocess_expressions_pathways}
\alias{preprocess_expressions_pathways}
\title{Formatting Gene Expression and Pathway Data to Use with Sparse Group Lasso}
\usage{
preprocess_expressions_pathways(
  gene_expr,
  all_pathways,
  phenotype_filename = NULL,
  phenotype_colname = NULL,
  covariates_filename = NULL,
  covariates_colnames = NULL,
  family_func = gaussian(link = "identity")
)
}
\arguments{
\item{gene_expr}{Dataframe with first 2 columns called FID and IID
containing unique participant IDs and columns named with gene names,
containing imputed gene expressions.}

\item{all_pathways}{List named by pathway, where each index contains a
vector of genes in the pathway.}

\item{phenotype_filename}{Path to text file with a column called FID
containing unique participant IDs and a column with the phenotype
of interest.}

\item{phenotype_colname}{Name of the column containing the phenotype of
interest in the phenotype file specified above.}

\item{covariates_filename}{Path to text file with a column called FID
containing unique participant IDs and columns with covariates.}

\item{covariates_colnames}{Vector of names of the columns containing the
covariates in the covariate file specified above.}

\item{family_func}{Family object specifying the type of model you
want to use in the sparse group lasso function. See ?stats::family for more
details.}
}
\value{
Returns a list(
X = gene expression matrix,
y = residuals on phenotype ~ covariates if covariates_filename != NULL
and family_func = gaussian(link = "identity"), or just vector of phenotype
data otherwise,
groups = vector of consecutive integers describing the pathway grouping of
the genes,
gene2pathway = list with genes as names and the
length of pathways that contain the gene,
FID = the IDs for the participants in order ,
offsets = NULL if the no covariates are specified or
family_func = gaussian(link = "identity"), vector of predicted values
from generalized linear regression using specified covariates otherwise,
processed_pathways = list with pathways as names and the genes in each
pathway formatted as gene_pathway).
}
\description{
Duplicates genes in your imputed gene expression matrix that appear multiple
times in your specified pathways, renames genes to gene_pathway, subsets and
reorders your gene expression and phenotype files to match the participant
ID order, and regresses out the covariates from your phenotype if you
specify family_func = gaussian(link = "identity") or provides offsets from
predicted values from glm(phenotype ~ ., data = covars, family = family_func).
}
