% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expr_heatmap_plot.R
\name{redim_matrix}
\alias{redim_matrix}
\title{Down Sample Matrix}
\usage{
redim_matrix(
  mat,
  cl = NULL,
  target_height = 300,
  target_width = 300,
  summary_func = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{mat}{Gene expression correlation matrix.}

\item{cl}{Parallel cluster.}

\item{target_height}{Max number of rows that the output matrix should have.}

\item{target_width}{Max number of columns that the output matrix should have.}

\item{summary_func}{Function with which to down sample the matrix.}
}
\value{
Returns a heatmap image of gene expression correlations.
}
\description{
Produces an a matrix of smaller dimensions from the input matrix, but with
similar patterns. Adapted from
https://gdevailly.netlify.app/post/plotting-big-matrices-in-r/ .
}
\references{
Devailly, G. (2021, October 14). Plotting heatmaps from big matrices
in R. \url{https://gdevailly.netlify.app/post/plotting-big-matrices-in-r/}.

R Core Team (2024). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria. https://www.R-project.org/.
}
