% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAS_to_adj_dataframe.R
\name{sgl2adj_df}
\alias{sgl2adj_df}
\title{Format Coefficient Output from Multiple Sparse Group Lassos into Adjacency Matrix}
\usage{
sgl2adj_df(gene_coefs, tissue_names = seq_along(gene_coefs))
}
\arguments{
\item{gene_coefs}{List of coefficient output from cv.sparsegl object.}

\item{tissue_names}{Names for the tissues/cell/condition types for the
imputed gene expression data used to train each cv.sparsegl object that
produced the above coefficients.}
}
\value{
Returns a list(
gene = gene adjacency-matrix-like data frame,
pathway = pathway adjacency-matrix-like data frame
).
}
\description{
Produces an adjacency-matrix-like dataframe where each row is a gene and
each column is the summed estimated effect size of the gene from a sparse
group lasso TWAS. Also produces another adjacency-matrix-like dataframe
but with pathways.
}
