% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot of Gene Effect Sizes vs. p-values}
\usage{
volcano_plot(
  betas_p_vals,
  p_thresh = 0.05,
  vline = c(qnorm(0.025), qnorm(0.975)),
  add_sig_gene_labels = TRUE
)
}
\arguments{
\item{betas_p_vals}{Dataframe with columns gene_name (to label points with),
betas (containing gene effect sizes), p_value (significance of effect sizes)}

\item{p_thresh}{Max value of p value to be considered significant.}

\item{vline}{Scalar or vector of values at which to draw vertical lines
on volcano plot.}

\item{add_sig_gene_labels}{Boolean. If TRUE, significant points will be
labeled with their gene names. If FALSE, no points will be labeled.}
}
\value{
Returns a ggplot volcano plot.
}
\description{
Produces volcano plot of gene effect sizes vs. -log10(p-values).
Labels significant genes.
}
\examples{
volcano_plot(betas_p_vals = data.frame(beta = c(1, 0, 3, 4, 0.5),
                                 p_value = c(0.1, 0.9, 0.02, 0.11, 0.3)))

}
