% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot of Gene Effect Sizes vs. p-values}
\usage{
volcano_plot(
  betas_p_vals,
  pvalue_colname = "pvalue",
  gene_colname = "gene",
  effect_size_colname = "zscore",
  p_thresh = 0.05,
  vline = c(qnorm(0.025), qnorm(0.975)),
  add_sig_gene_labels = TRUE
)
}
\arguments{
\item{betas_p_vals}{Data frame with effect sizes and p-values.}

\item{pvalue_colname}{Name of column in predixcan_assoc_filenames that
contains the p-value.}

\item{gene_colname}{Name of column in predixcan_assoc_filenames that contains
the gene name.}

\item{effect_size_colname}{Name of column in predixcan_assoc_filenames that
contains the effect size.}

\item{p_thresh}{Numeric. P-value threshold for significance (default = 0.05).}

\item{vline}{Numeric vector. Vertical reference lines (default uses ±1.96).}

\item{add_sig_gene_labels}{Logical. Whether to label significant genes.}
}
\value{
Returns a ggplot volcano plot.
}
\description{
Produces volcano plot of gene effect sizes vs. -log10(p-values).
Labels significant genes.
Meant to take in TWAS results from PrediXcan which contain a gene, zscore
(standardized effect size), and a pvalue column, summarizing the individual
associations each gene has with your phenotype of interest.
}
\examples{
volcano_plot(betas_p_vals = data.frame(zscore = c(1, 0, 3, 4, 0.5),
                                 pvalue = c(0.1, 0.9, 0.02, 0.11, 0.3)))

}
